Module rust_lib_flutter_test_gui::frb_generated

Re-exports

pub use io::*;

Modules

io 🔒

Structs

FLUTTER_RUST_BRIDGE_HANDLER

IdGenerator 🔒

MoiArc

MoiArcPoolInner

MoiArcPoolValue 🔒

StreamSink

Constants

FLUTTER_RUST_BRIDGE_CODEGEN_CONTENT_HASH 🔒

FLUTTER_RUST_BRIDGE_CODEGEN_VERSION 🔒

Traits

CstDecode

MoiArcValue

SseDecode

SseEncode

Functions

decode_rust_opaque_moi 🔒

pde_ffi_dispatcher_primary_impl 🔒

pde_ffi_dispatcher_sync_impl 🔒

transform_result_sse 🔒

wire_generate_random_name_impl 🔒

wire_init_app_impl 🔒

wire_start_rust_receiver_impl 🔒

wire_start_rust_sender_impl 🔒

Type Aliases

MoiArcPool

ObjectId 🔒

RustAutoOpaque

Usually this is unneeded, and just write down arbitrary types. However, when you need arbitrary types at places that are not supported yet, use RustOpaqueOpaque<YourArbitraryType>.

RustAutoOpaqueMoi

Please refer to RustAutoOpaque for doc.

RustOpaque

A wrapper to support arbitrary Rust types.

RustOpaqueMoi

Please refer to RustOpaque for doc.